requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-okx==0.3.9
redis
python-okx==0.3.9
redis
pandas==2.1.3
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
